modularize ser/de
